+++ packages/slate-plugins/src/common/plugins/index.ts
@@ -1,3 +1,3 @@
 export * from './inline-void';
 export * from './node-id';
-export * from './withToggleType';
+export * from './toggle-type';