+++ packages/slate-plugins/src/elements/table/index.ts
@@ -1,7 +1,6 @@
 export * from './components';
 export * from './defaults';
 export * from './deserializeTable';
-export * from './queries';
 export * from './renderElementTable';
 export * from './TablePlugin';
 export * from './types';