+++ packages/slate-plugins/src/elements/table/queries/index.ts
@@ -1,3 +1,0 @@
-export * from './isTable';
-export * from './isTableCell';
-export * from './isTableRow';