+++ packages/slate-plugins/src/elements/table/transforms/addColumn.ts
@@ -1,21 +1,18 @@
 import { Editor, Path, Transforms } from 'slate';
-import { isNodeTypeIn } from '../../../common/queries';
+import { getAboveByType, isNodeTypeIn } from '../../../common/queries';
 import { setDefaults } from '../../../common/utils/setDefaults';
 import { DEFAULTS_TABLE } from '../defaults';
-import { isTable, isTableCell } from '../queries';
 import { TableOptions } from '../types';
 import { getEmptyCellNode } from '../utils';
 
 export const addColumn = (editor: Editor, options?: TableOptions) => {
-  const { table } = setDefaults(options, DEFAULTS_TABLE);
+  const { table, td, th } = setDefaults(options, DEFAULTS_TABLE);
 
   if (isNodeTypeIn(editor, table.type)) {
-    const currentCellItem = Editor.above(editor, {
-      match: isTableCell(options),
-    });
-    const currentTableItem = Editor.above(editor, {
-      match: isTable(options),
-    });
+    const currentCellItem = getAboveByType(editor, [td.type, th.type]);
+
+    const currentTableItem = getAboveByType(editor, table.type);
+
     if (currentCellItem && currentTableItem) {
       const nextCellPath = Path.next(currentCellItem[1]);
       const newCellPath = nextCellPath.slice();