+++ packages/slate-plugins/src/handlers/reset-block-type/index.ts
@@ -1,4 +1,3 @@
+export * from './onKeyDownResetBlockType';
+export * from './ResetBlockTypePlugin';
 export * from './types';
-export * from './withBreakEmptyReset';
-export * from './withDeleteStartReset';
-export * from './withResetBlockType';